+++ examples/Tutorial_1-Document_Indexing/Step_2_IndexStories.sh
@@ -16,7 +16,8 @@
 # 3. The index is stored in the specified path "./full_text_index".
 ###################################################################################
 
-chatterlang_script --script "
+#chatterlang_script --script "
+python -m talkpipe.app.runscript --script "
     INPUT FROM \"stories.json\" 
     | readJsonl 
     | progressTicks[tick_count=1, print_count=True]