+++ examples/Tutorial_2-Search_by_Example_and_RAG/Step_2_SearchByExample.sh
@@ -16,7 +16,7 @@
 ###################################################################################
 
 #talkpipe_endpoint --form-config story_by_example_ui.yaml --script "
-python -m talkpipe.app.apiendpoint --form-config story_by_example_ui.yaml --script "
+python -m talkpipe.app.apiendpoint --form-config story_by_example_ui.yaml --display-property example --script "
     | copy
     | llmEmbed[field=\"example\", source=\"ollama\", model=\"mxbai-embed-large\", append_as=\"vector\"]
     | searchVector[vector_field=\"vector\", path=\"./vector_index\"]